User context should log when users are deleted.
There are two hooks on wp_delete_user: `delete_user` and `deleted_user`.
Logging happens on `deleted_user` action since it will be invoked when
user is successfully removed from DB. The param $user_id received in
`deleted_user` is useless as the records was already removed from DB, so
that to we need to hook into `delete_user` to store user object temporarily.

This fixes issue #32.