Improving theme data gathering on Update On Issue #241 @fjrrett reported some notices on the theme update page.
The source of the problem there was that the information can be stored
in another variable called `theme_info` and it is a instance of [WP
Theme](http://codex.wordpress.org/Class_Reference/WP_Theme) which the
methods change a bit from what was used.

I've included a conditional to check for the `theme` variable them if is
not set I work with the `WP Theme` instance.