Prevent UI from allowing invalid date intervals
Fixes issue #292.

I chose not to set readonly to the text inputs in the HTML so that
these fields can still work without javascript enabled.

If I were to set readonly to the datepicker inputs at document load and
mimic a clickable input through styling, I would have to override the
WordPress admin’ input[readonly]’ style with
    background-color: #fff;
But this assumes that clickable inputs will always have a background
colour of white.