Enable notifications network wide (read description)
Important! In order for notifications to work network wide, we need to
be storing the notification meta data with update_site_option, not
update_option.

This means all notification meta will be stored in the same table on a
multisite install. This lets us easily access meta data for network
notifications, even when we’re not in the network admin.

This will mean that any existing rules that you have set up WILL BREAK
with this change. You will need to set them up again.

I could create a fallback, but since this hasn’t been released yet, I
don’t think it’s necessary.